% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/botcheck.R
\name{botcheck}
\alias{botcheck}
\title{botcheck}
\usage{
botcheck(user)
}
\arguments{
\item{user}{The Twitter handle you want to check, e.g., "barackobama". The "@" is not required.}
}
\description{
This function makes a call to the Botomter API and returns the probability that a specified Twitter user is a bot.
}
\examples{
botcheck("barackobama")
}
\keyword{API}
\keyword{Botometer}
\keyword{wrapper}
